Changed logic to pull data instead of hardcode
This took literally all day because I'm too stubborn.

I changed the logic to use actual data instead of hardcoded values. I chose to use an ajax method as that's what I'm familiar with at work. It gets the data, builds the content and displays the results, and always shows updated content with no need to refresh. Not that this API will ever get updated...

Added an error condition during testing so that it's display is blank if nothing is returned. I had a LOT of issues getting it to display correctly, thus console messages.

Removed some unused CSS